Simple yaml file for internal use
I think this will work for all projects
not just ones that will have info files.

This file can have the top level values
It can answer questions like where is the code.
Do we use jenkins, gerrit, etc.

When creating a new repository or jira for a project,
If that project is not providing an info file, this could be
part of the base information we need.

It could be used in templating...

I need to know more about how non gerrit projects operate.

And I need to talk to whoever is working on terraform.

Change-Id: Icfd35a72af4cb2e782b790aebb598fefe637175a
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>