Simple yaml file for internal use
Lists projects and their respective release engineers
will be used when adding reviewers for info file changes.

We can expland and change this as needed when we find more uses.
(patches welcome)

My current half baked plan is:

we should add all the projects not just ones that will have info files.
we then have the values that are normally in the info file for jira etc.
for project/repo creation etc.
this file, I think, will be the main driver for project creation

Also I just best guessed on the values.
so something is probably wrong.

Change-Id: Icfd35a72af4cb2e782b790aebb598fefe637175a
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>