Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Morgan, 
> 
> Can we double check the committers in this list. They might be ok, but just wanted to make sure we want to add them. I think some of them are no longer active committers anymore.. is it?

the list is based on the current up to date Integration list
It is right that some people are not so active
I could clean a little bit but except for Catherine (not really contributing but historically declared as Integration committers) I planned to give 2 releases grace delay before cleaning.
I did that for Gary.
I planned to do it for Brian, Daniel and Marco (so clean them on Honolulu as their last contributions is from Frankfurt), in case they would come back..it is a manual hysteresis even if i doubt they will come back
the last one is not an integration committer, he is the main developer of the code for this repo.
I give him a +2/merge only on this repo, and I am responsible as integration PTL if he starts doing things crazy on this particular repo

Patch-set: 1