Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> Please submit this change to the oran $projects INFO.yaml file. Should not be submitted directly to info-master repo.

This comment is not clear to me. There was a committer added in the INFO.yaml file in info-master repo, and I was just adding another committer. Are you saying that the INFO.yaml in info-master should not have had a committer to begin with? If so, could someone point me to a sample of INFO.yaml file at a project level?

Patch-set: 1