Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> @Dan, 
> 
> The issue is that in Gerrit there is already a parent repo called "ccsdk/sli" This repo is hidden since it was created in the beginning of the project as a parent for all the repos under ccsdk/sli/*
> 
> Can we name this project something else? or create it under ccsdk/sli/<name>

I really would prefer not to rename (the groupIds in our code have to match the repo structure .. so if we change to a new name, we'd have to change all our groupIds in all our poms, which would be a breaking change for any clients using those repos). Is there a way to just unhide that existing repo so we can just use that?

Patch-set: 1