Update patch set 2
Patch Set 2:

Hi Morgan, Sorry for the inconvenience, but could we please cancel this submission? Our original proposed repo name “transport-network-controller-simulator” may not be appropriate because our simulator uses the IETF/ACTN standards, which is not the only standard to be used for the transport network controllers. (There are other standards, such as ONF/TAPI, which can also be used for controllers.) So the original name may be misleading or cause confusion. 

We propose to create a new repo named "integration/ieft-actn-tools", which contains several tools used for transport slicing integration test, including the transport network controller simulator. We can review this proposal in the weekly integration meetings. 

Thanks.

Patch-set: 2
Reviewer: Gerrit User 527 <527@f2a2391d-0403-4fa5-b183-4056999a8a09>